set up the API /DTO /DAO

+++ Recrutement-CV-Service/src/main/java/org/tarik/GestionRecrutement/dao/CandidatRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.tarik.GestionRecrutement.model.Candidat;
 @Repository
-public interface CandidatRepository extends  JpaRepository<Candidat, Integer>{
+public interface CandidatRepository extends  JpaRepository<Candidat, Long>{
 
 }
